Enhancement: Add File Dialog to Browse for Images
Add the ability to use a file browse dialog to open a file from another folder
on disk using Ctrl+O / F1. This is useful when testing or launching the tool
from a shortcut, where it may not be started with a filename.

This makes use of the "nfd" crate (which uses the "nativedialog" library
under the hood, requiring a working C/C++ compiler for compilation to work)

Limitations:
* The file dialog allows selecting any file types, and not just supported
  image file formats.