Gelatin Window - Add a .maximized() flag
Add a flag + getter() for checking the window is maximised (similar to the
`fullscreen` stuff), so that this state can be accurately restored as part
of the cache too.

This is part of the fix for window sizing + positioning problems when on
startup if a previous session was used maximized.