Fix: Enforce a minimum width for the slider indicator
On folders with > 1000 images (as tested with a folder with 3-4k images),
the slider indicator width could get set to such small values that it
wasn't visible anymore. Now, clamp to a minimum thickness so that it is
always sufficiently visible.

Note: Couldn't use `cmp::max()`, due to Rust's stupid pedantry around
      float comparisons