Window Size/Position Bugfix - Part 2
Introduce some hardcoded fudge factors to account for the Windows taskbar
reducing the available screen area. This hack only really works correct
when the taskbar is in the default position (on the bottom of the screen).

This should be "good enough" for the general case (i.e. my use case,
and that of the other guy who reported this on the original / upstream repo), but ultimately, a proper cross-platform solution should be found.

For reference of the bug this fixes, see https://github.com/ArturKovacs/emulsion/issues/210